+++ tests/unit/sync/io/test_neo4j_pool.py
@@ -32,7 +32,7 @@ from neo4j.conf import (
     WorkspaceConfig,
 )
 
-from ..._async_compat import (
+from ...._async_compat import (
     mark_sync_test,
     Mock,
 )