+++ tests/unit/sync/test_driver.py
@@ -22,13 +22,11 @@ from neo4j import (
     BoltDriver,
     GraphDatabase,
     Neo4jDriver,
-    TRUST_ALL_CERTIFICATES,
-    TRUST_SYSTEM_CA_SIGNED_CERTIFICATES,
 )
 from neo4j.api import WRITE_ACCESS
 from neo4j.exceptions import ConfigurationError
 
-from .._async_compat import (
+from ..._async_compat import (
     mark_sync_test,
     mock,
 )
@@ -70,15 +68,15 @@ def test_routing_driver_constructor(protocol, host, port, auth_token):
         ({"encrypted": False}, ConfigurationError, "The config settings"),
         ({"encrypted": True}, ConfigurationError, "The config settings"),
         (
-            {"encrypted": True, "trust": TRUST_ALL_CERTIFICATES},
+            {"encrypted": True, "trusted_certificates": []},
             ConfigurationError, "The config settings"
         ),
         (
-            {"trust": TRUST_ALL_CERTIFICATES},
+            {"trusted_certificates": []},
             ConfigurationError, "The config settings"
         ),
         (
-            {"trust": TRUST_SYSTEM_CA_SIGNED_CERTIFICATES},
+            {"trusted_certificates": None},
             ConfigurationError, "The config settings"
         ),
     )
@@ -103,21 +101,6 @@ def test_driver_config_error(
 def test_invalid_protocol(test_uri):
     with pytest.raises(ConfigurationError, match="scheme"):
         GraphDatabase.driver(test_uri)
-
-
-@pytest.mark.parametrize(
-    ("test_config", "expected_failure", "expected_failure_message"),
-    (
-        ({"trust": 1}, ConfigurationError, "The config setting `trust`"),
-        ({"trust": True}, ConfigurationError, "The config setting `trust`"),
-        ({"trust": None}, ConfigurationError, "The config setting `trust`"),
-    )
-)
-def test_driver_trust_config_error(
-    test_config, expected_failure, expected_failure_message
-):
-    with pytest.raises(expected_failure, match=expected_failure_message):
-        GraphDatabase.driver("bolt://127.0.0.1:9001", **test_config)
 
 
 @pytest.mark.parametrize("uri", (