+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Neo4j Driver Change Log
+# Neo4j Driver Change Log (breaking/major changes only)
 
 ## Version 5.0
 
@@ -8,6 +8,13 @@
   `neo4j.work`. They should've been imported from `neo4j` all along.
 - Experimental pipelines feature has been removed.
 - Experimental async driver has been added.
+- `ResultSummary.server.version_info` has been removed.  
+  Use `ResultSummary.server.agent`, `ResultSummary.server.protocol_version`,
+  or call the `dbms.components` procedure instead.
+- SSL configuration options have been changed:
+  - `trust` has been removed.  
+    Use `trusted_certificates` instead which expects `None` or a `list`. See the
+    API documentation for more details.
 - `neo4j.time` module:
  - `Duration`
    - The constructor does not accept `subseconds` anymore.  
@@ -42,6 +49,7 @@
      Use `hour_minute_second_nanosecond` instead.
    - The property `second` returns an `int` instead of a `float`.  
      Use `nanosecond` to get the sub-second information.
+
 
 ## Version 4.4